Fix integration tests
This changes the protocol used by byzcoin for the view change to be
the blsftcosi one. It also fixes multiple issues in the integration
tests.
