+++ identity/service.go
@@ -18,9 +18,9 @@ import (
 
 	"github.com/dedis/cothority"
 	"github.com/dedis/cothority/messaging"
+	"github.com/dedis/cothority/pop/service"
 	"github.com/dedis/cothority/skipchain"
 	"github.com/dedis/kyber"
-	"github.com/dedis/kyber/pairing"
 	"github.com/dedis/kyber/sign/anon"
 	"github.com/dedis/kyber/sign/schnorr"
 	"github.com/dedis/kyber/util/key"
@@ -49,7 +49,7 @@ var VerificationIdentity = []skipchain.VerifierID{skipchain.VerifyBase, VerifyId
 var VerifyIdentity = skipchain.VerifierID(uuid.NewV5(uuid.NamespaceURL, "Identity"))
 
 func init() {
-	identityService, _ = onet.RegisterNewServiceWithSuite(ServiceName, pairing.NewSuiteBn256(), newIdentityService)
+	identityService, _ = onet.RegisterNewService(ServiceName, newIdentityService)
 	network.RegisterMessage(&storage1{})
 	network.RegisterMessage(&IDBlock{})
 }
@@ -111,7 +111,8 @@ func (s *Service) StoreKeys(req *StoreKeys) (network.Message, error) {
 				"Invalid request")
 
 		}
-		if err := req.Final.VerifyWithService(ServiceName); err != nil {
+
+		if err := req.Final.VerifyWithService(service.Name); err != nil {
 			log.Error(s.ServerIdentity(), "Invalid FinalStatement: ", err)
 			return nil, errors.New(
 				"Signature of final statement is invalid: " + err.Error())
